@vilkris4 2024-08-25
After further testing I have [removed](https://github.com/vilkris4/go-zenon/commit/844f69aba35757c0763f46cbaa32f581944bbf1e) the changes done to the account pool's rebuild functionality from this PR because of the following reason (copy of the commit message):

> In a situation where an account chain continuously has uncommitted blocks in the account pool, the "restabilization" of the `memdbManager` will eventually result in a massive amount of wrapped databases inside wrapped databases inside wrapped databases in the `memdbManager`'s `versions` map. This is because the `versions` map never gets fully cleared and rebuilt from scratch when the `Restabilize` method is used. These wrapped databases inside wrapped databases can cause significant issues in read performance when reading from the unconfirmed account database.
> 
> Because of this the changes to the rebuild functionality have been reverted, even though this issue is only an edge case. More consideration is needed on how to improve the account pool's rebuild performance, while minimizing unwanted side effects.