@0x3639 2024-09-19
For anyone interested in researching this PR here is a two part code review between George and Vilkris.  

Part 1/2
https://youtu.be/Pb87qKrtszs

Part 2/2
https://youtu.be/CZxMgTuKSuk